+++ src/main/java/org/acme/domain/dtos/SimulationsResultDto.java
@@ -1,11 +1,12 @@
 package org.acme.domain.dtos;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 public record SimulationsResultDto(
         Long simulationId,
         Long productId,
         String productName,
-        Double interestRate,
+        BigDecimal interestRate,
         ArrayList<PresentSimulationDto> simulationResult
 ) { }
